+++ src/main/java/com/arangodb/entity/ErrorEntity.java
@@ -26,7 +26,7 @@ import java.io.Serializable;
  * @author Mark Vollmary
  *
  */
-public class ErrorEntity implements Serializable {
+public class ErrorEntity implements Serializable, Entity {
 
 	private static final long serialVersionUID = -5918898261563691261L;